# backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  title: NomNomVault API
  version: 0.1.0
paths:
...
  "/media/{mediaId}/original":
    get:
      description: Stream the original media bytes for the requested asset.
      parameters:
        - description: Media ID
          in: path
          name: mediaId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: Fetch recipe media
      tags:
        - recipes
  /recipes:
    get:
      description: Return the current household recipe list using cursor pagination,
        search, and tag filters.
      parameters:
        - description: Cursor token
          in: query
          name: cursor
          schema:
            type: string
        - description: Maximum number of recipes to return
          in: query
          name: limit
          schema:
            type: integer
        - description: Case-insensitive recipe search query
          in: query
          name: q
          schema:
            type: string
        - description: Repeatable tag slug filters
          in: query
          name: tag
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpapi.RecipeListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: List recipes
      tags:
        - recipes
    post:
      description: Create a recipe for the active household.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/httpapi.CreateRecipeRequest"
        description: Create recipe payload
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpapi.RecipeDetailResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: Create a recipe
      tags:
        - recipes
  "/recipes/{recipeId}":
    get:
      description: Return the detailed recipe payload for the requested recipe ID.
      parameters:
        - description: Recipe ID
          in: path
          name: recipeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpapi.RecipeDetailResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: Fetch a recipe detail
      tags:
        - recipes
    patch:
      description: Replace the editable recipe fields and child collections for the
        active household.
      parameters:
        - description: Recipe ID
          in: path
          name: recipeId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/httpapi.UpdateRecipeRequest"
        description: Update recipe payload
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpapi.RecipeDetailResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: Update a recipe
      tags:
        - recipes
  "/recipes/{recipeId}/archive":
    post:
      description: Mark a recipe as archived for the active household.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "403":
          description: Forbidden
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: Archive a recipe
      tags:
        - recipes
  "/recipes/{recipeId}/media":
    post:
      description: Upload an image and attach it to the requested recipe.
      parameters:
        - description: Recipe ID
          in: path
          name: recipeId
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: Recipe image file
                  type: string
                  format: binary
                altText:
                  description: Optional alt text
                  type: string
              required:
                - file
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpapi.RecipeMediaItem"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: Upload recipe media
      tags:
        - recipes
  "/recipes/{recipeId}/unarchive":
    post:
      description: Mark an archived recipe as active again for the active household.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "403":
          description: Forbidden
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/apicontract.ErrorResponse"
      summary: Unarchive a recipe
      tags:
        - recipes
...
servers:
  - url: /api/v1
components:
  schemas:
...
    apicontract.ErrorResponse:
      properties:
        code:
          type: string
        details:
          items:
            $ref: "#/components/schemas/apicontract.ValidationError"
          type: array
        message:
          type: string
        requestId:
          type: string
        status:
          type: integer
      type: object
    apicontract.ValidationError:
      properties:
        field:
          type: string
        message:
          type: string
      type: object
...
    httpapi.CreateRecipeRequest:
      properties:
        cookMinutes:
          type: integer
        description:
          type: string
        ingredients:
          items:
            $ref: "#/components/schemas/httpapi.CreateRecipeIngredientRequest"
          type: array
        prepMinutes:
          type: integer
        servings:
          type: integer
        sourceUrl:
          type: string
        title:
          type: string
      required:
        - ingredients
        - title
      type: object
...
    httpapi.RecipeDetailResponse:
      properties:
        ingredients:
          items:
            $ref: "#/components/schemas/httpapi.RecipeIngredientItem"
          type: array
        mediaAssets:
          items:
            $ref: "#/components/schemas/httpapi.RecipeMediaItem"
          type: array
        nutritionEntries:
          items:
            $ref: "#/components/schemas/httpapi.RecipeNutritionItem"
          type: array
        recipe:
          $ref: "#/components/schemas/httpapi.RecipeSummary"
        steps:
          items:
            $ref: "#/components/schemas/httpapi.RecipeStepItem"
          type: array
        tags:
          items:
            $ref: "#/components/schemas/httpapi.RecipeTagItem"
          type: array
      type: object
...
    httpapi.RecipeListResponse:
      properties:
        data:
          items:
            $ref: "#/components/schemas/httpapi.RecipeSummary"
          type: array
        page:
          $ref: "#/components/schemas/apicontract.CursorPageInfo"
      type: object
...
    httpapi.UpdateRecipeRequest:
      properties:
        cookMinutes:
          type: integer
        description:
          type: string
        ingredients:
          items:
            $ref: "#/components/schemas/httpapi.CreateRecipeIngredientRequest"
          type: array
        nutritionEntries:
          items:
            $ref: "#/components/schemas/httpapi.UpdateRecipeNutritionRequest"
          type: array
        prepMinutes:
          type: integer
        servings:
          type: integer
        sourceUrl:
          type: string
        steps:
          items:
            $ref: "#/components/schemas/httpapi.UpdateRecipeStepRequest"
          type: array
        tagIds:
          items:
            type: string
          type: array
        title:
          type: string
        version:
          type: integer
      required:
        - ingredients
        - title
        - version
      type: object
...